add format as alias for span and move string_to_color to standard_colors.cpp
