+++ src/help/help_text_area.cpp
@@ -18,6 +18,7 @@
 #include "config.hpp"                   // for config, etc
 #include "draw.hpp"                     // for blit, fill
 #include "font/sdl_ttf_compat.hpp"
+#include "font/standard_colors.hpp"     // for string_to_color
 #include "game_config.hpp"              // for debug
 #include "help/help_impl.hpp"           // for parse_error, box_width, etc
 #include "lexical_cast.hpp"
@@ -294,7 +295,7 @@ void help_text_area::handle_format_cfg(const config &cfg)
 	bool bold = cfg["bold"].to_bool();
 	bool italic = cfg["italic"].to_bool();
 	int font_size = cfg["font_size"].to_int(normal_font_size);
-	color_t color = help::string_to_color(cfg["color"]);
+	color_t color = font::string_to_color(cfg["color"]);
 	add_text_item(text, "", false, font_size, bold, italic, color);
 }
